Fix inneficiency when output is full
Check to see if we can actually put stuff into the output inventory
before checking each tree.

There are two potential ways to remove the hard-coded value.  First, we
could look up the max amount that can drop from a tree at runtime and
put it there.  Alternatively, we could allow the output stack to go over
max if it was under max before we tried to add stuff to it.  This is
already done in some places for the vanilla game.