Cache the known stack sizes
On a test map with 31 full farms, I couldn't get the UPS to drop below
60.  Before this change, it was hovering between 1.5-2.5 ms spent in
this mod. After this change, it was hovering around 0.15-0.2 ms, with
occasional spikes up to 0.5 ms.
When using only Matrix0150's change on the same save, I get 1.25-2.25 ms
spent in TF.  When combining this change with Matrix0150's change, it
drops to 0.125-0.2 on average with spikes at 0.25.
This patch to my change is certainly the fastest performing fix for this
performance problem, but it appears to smoothed out by Matrix0150's
change, so combining them would be ideal.